Add live solutions to LeetCode Weekly 342
If you see this, message me with some tips on how to solve the last
problem more quickly. I took forever to come up with this idea but the
idea was frankly fairly simple. I just wasn't drawn to think of it for
mysterious reasons.

Also, if LeetCode continues to include trivial and almost silly problems
like A and B (and A through C from last contest), I will not attend
their contests anymore.